RAG Application with Groq/requirements.txt
flask==3.0.2
faiss-cpu==1.8.0
sentence-transformers==2.7.0
numpy==1.26.4
requests==2.31.0
torch>=2.2.0
transformers>=4.41.0
scikit-learn>=1.3.0
scipy>=1.10.0